refactor: rename ErrorMesages class
